+++ infra/resources/_modules/cosmos/cosmos_containers.tf
@@ -1,18 +1,59 @@
-module "wallet_cosmosdb_containers" {
+resource "azurerm_cosmosdb_sql_container" "containers_01" {
   for_each = { for c in local.wallet_cosmosdb_containers : c.name => c }
-
-  source = "github.com/pagopa/terraform-azurerm-v3//cosmosdb_sql_container?ref=v8.12.2"
 
   name                = each.value.name
   resource_group_name = var.resource_group_name
 
-  account_name  = azurerm_cosmosdb_account.wallet.name
-  database_name = azurerm_cosmosdb_sql_database.db.name
-
+  account_name       = azurerm_cosmosdb_account.wallet.name
+  database_name      = azurerm_cosmosdb_sql_database.db.name
   partition_key_path = each.value.partition_key_path
-  throughput         = lookup(each.value, "throughput", null)
+  default_ttl        = each.value.default_ttl
 
-  autoscale_settings = lookup(each.value, "autoscale_settings", null)
+  autoscale_settings {
+    max_throughput = each.value.autoscale_settings.max_throughput
+  }
+}
 
-  default_ttl = lookup(each.value, "default_ttl", null)
+resource "azurerm_cosmosdb_sql_container" "leases_01" {
+  name                = "leases"
+  resource_group_name = var.resource_group_name
+
+  account_name       = azurerm_cosmosdb_account.wallet.name
+  database_name      = azurerm_cosmosdb_sql_database.db.name
+  partition_key_path = "/id"
+  default_ttl        = null
+
+  autoscale_settings {
+    max_throughput = 1000
+  }
+}
+
+resource "azurerm_cosmosdb_sql_container" "containers_02" {
+  for_each = { for c in local.wallet_cosmosdb_containers : c.name => c }
+
+  name                = each.value.name
+  resource_group_name = var.resource_group_name
+
+  account_name       = local.cosmos_02.name
+  database_name      = azurerm_cosmosdb_sql_database.db.name
+  partition_key_path = each.value.partition_key_path
+  default_ttl        = each.value.default_ttl
+
+  autoscale_settings {
+    max_throughput = each.value.autoscale_settings.max_throughput
+  }
+}
+
+resource "azurerm_cosmosdb_sql_container" "leases_02" {
+  name                = "leases-revoke-wallet-instance"
+  resource_group_name = var.resource_group_name
+
+  account_name       = local.cosmos_02.name
+  database_name      = azurerm_cosmosdb_sql_database.db.name
+  partition_key_path = "/id"
+  default_ttl        = null
+
+  autoscale_settings {
+    max_throughput = 1000
+  }
 }