+++ infra/resources/_modules/cosmos/locals.tf
@@ -6,7 +6,7 @@ locals {
       name               = "wallet-instances"
       partition_key_path = "/userId"
       autoscale_settings = {
-        max_throughput = 20000
+        max_throughput = 4000
       }
       default_ttl = null
     },
@@ -15,18 +15,15 @@ locals {
       name               = "nonces"
       partition_key_path = "/id"
       autoscale_settings = {
-        max_throughput = 15000
+        max_throughput = 4000
       }
       default_ttl = 300
     },
-    # Leases for trigger
-    {
-      name               = "leases"
-      partition_key_path = "/id"
-      autoscale_settings = {
-        max_throughput = 1000
-      }
-      default_ttl = null
-    }
   ]
+
+  # Temporary, see CES-535
+  cosmos_02 = {
+    id   = "/subscriptions/ec285037-c673-4f58-b594-d7c480da4e8b/resourceGroups/io-p-itn-wallet-rg-01/providers/Microsoft.DocumentDB/databaseAccounts/io-p-itn-wallet-cosno-02"
+    name = "io-p-itn-wallet-cosno-02"
+  }
 }