+++ infra/resources/_modules/ids/variables.tf
@@ -1,0 +1,19 @@
+variable "project" {
+  type        = string
+  description = "IO prefix and short environment"
+}
+
+variable "location" {
+  type        = string
+  description = "Azure region"
+}
+
+variable "tags" {
+  type        = map(any)
+  description = "Resource tags"
+}
+
+variable "resource_group_name" {
+  type        = string
+  description = "Name of the resource group where resources will be created"
+}